llm-service/pyproject.toml
[project]
name = "llm-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.11",
    "langchain-ollama>=0.2.3",
    "langgraph>=0.3.18",
    "python-dotenv>=1.0.1",
    "tavily-python>=0.5.1",
]
